Fix warning in DefaultFormatter

+++ src/formatter.cpp
@@ -32,7 +32,7 @@ Buffer DefaultFormatter::format(const Record& record) const
 
     auto currentTime = (record.steadyTp - record.startTp);
     unsigned long long secs = duration_cast<seconds>(currentTime).count();
-    uint16_t msecs = duration_cast<milliseconds>(currentTime).count() - secs * 1000;
+    uint16_t msecs = static_cast<uint16_t>(duration_cast<milliseconds>(currentTime).count() - secs * 1000);
 
     ALog::LongSSO<> result(impl().cache);
     result.appendFmtString("[%5llu.%03hu] T#%-2d ", secs, msecs, record.threadNum);
